Modify Bitcoin Client
Add possibility to simulate mining. Start the client with -simuLambda=n to launch bitcoind in simulated mode, n is the desired default lambda.
If simuLamda is not present when initializing the application, real mining will be used.

Add possibility to specify default mining difficulty when client is launched in real mining mode.
Launching bitcoind with -dificulta=n will set n as the default mining dificulty. n specifies the amount of zeroes the hash has to have as a prefix

Add possibility to specify desired difficulty. Mine blocks as 'generate x n', x is the amount of blocks to mine, n is the desired difficulty.
In real mining mode n denotes the amount of zeroes the hash has to begin with.
In simulated mode n specifies the lambda to be utilized for the generation of the random value.

The client will no longer check the difficulty of incoming blocks from peers.
This allows the creation of heterogeneous networks, in which some nodes are doing real mining while other are in simulated mode

The Client will log in thesis_log_file.txt the creation of new blocks and the appearence of new block hashes sent from their peers

Add testscript.go, which is a script that instantiates nodes and makes them mine new blocks every a couple of seconds in order to test new changes